cast double to int (there will be no lossy conversion here since the maximum amt levels it would spit out is 30, with an xp amt well in int range)

+++ src/main/java/tb/common/tile/TileOverchanter.java
@@ -94,7 +94,7 @@ public class TileOverchanter extends TileEntity implements IInventory, IWandable
                                         .getBoundingBox(xCoord, yCoord, zCoord, xCoord + 1, yCoord + 1, zCoord + 1)
                                         .expand(6, 3, 6));
                                 if (!players.isEmpty()) {
-                                    int lvlsleft = Math.ceil(xpToAbsorb > 255 ?
+                                    int lvlsleft = (int) Math.ceil(xpToAbsorb > 255 ?
                                         (59 + Math.sqrt(24 * xpToAbsorb - 5159)) / 6 :
                                         xpToAbsorb / 17d);
                                     // it's a double in the second branch so that both branches use the same Math.sqrt
